Add cross reference processing script.
I wrote a new script to attempt parsing of cross references so that
proper osisRef attributes can be added to the reference tags.

It is experimental. It may not work for you. It does, however, work
for me on the one bible I had access to that included cross references.

More testing of this is needed.