Appease those who have made false accusations.
I recently became aware of some false claims about this code and
some false accusations about me in particular. So to appease those
who made those false accuastions I have embedded the xml.xsd schema
in the script and made some alterations in order to make sure that
there is absolutely NO network activity going on at any time during
the script's execution. Testing without networking has confirmed
that networking is not needed for this script to either process the
usfm files or validate the converted osis.

There are some other minor bug fixes and pylint issues addressed
as well.

On a completely separate note, why are christians so quick to assume
the worst about people? Aren't we supposed to be better than this?