always validate output.
I can't think of any reason why disabling validation would make sense. Nor can I think of any reason why disabling the reformatting of the osis xml makes sense. So I have adjusted the code so that it will always validate and reformat the output when the lxml module is present.

If you absolutely must have the messy xml that this script produces without running it through lxml, that is still available. You will have to use the -d option to get it though. Note that this is NOT recommended. You will also lose the multiprocessing built into this script. Debug is meant for debugging problems in the script itself, not the xml that it produces.

If there is a valid reason why this change should not be implemented, and evidence can be shown that it does something wrong, then I will revert this change.